PanDA/NoSQL: handle reshuffling of Cassandra tables
Deleted tables/models:

 - day_site_errors, day_site_errors_cnt_30m: these tables allow
   only midnight-boundary queries and we have day_mtime_* tables
   that allow arbitrary time range filters.

Reworked tables/models:

 - day_mtime_site_errors_cnt_30m was assimilated by day_site_errors_cnt:
   * the table will keep data for different intervals in one place;
   * all summary tables will have mtime, so there is no need to embed
     this knowledge into their names.

Since now all index tables support non-midnight boundaries, there is
no need to have flag 'base_mtime range query?'.

And now we run queries which include both dates and modification time:
this simplifies the code and testing (we will always have same query
type).  Since the default is to query last 3 hours and people (sites)
can be interested in errors for past 12h, 24h, 3h and not just for the
whole day, letting the old code to be present would be
a micro-optimization that won't be used in majority of cases.

This commit restricts ourselves just to '30m' bins processing to mimic
the previously existing logics.  Handling of other bin types is on its
way, so stay tuned.

Signed-off-by: Eygene Ryabinkin <rea@grid.kiae.ru>